% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{University Graduate Income}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480 rows and 7 columns.}
\source{
Statistics Canada. Table 37-10-0115-01: Characteristics and median
employment income of longitudinal cohorts of postsecondary graduates two and
five years after graduation, by educational qualification and field of study,
2010 and 2011 cohorts. \url{https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=3710011501}
}
\usage{
income
}
\description{
University Graduate Income
}
\keyword{datasets}
