% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment}
\alias{enrollment}
\title{Canadian University Enrollment}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100786 rows and 9 columns.}
\source{
Table: 37-10-0011-01 (formerly: CANSIM 477-0029)
\url{https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=3710001101}
}
\usage{
enrollment
}
\description{
Canadian University Enrollment by year, province,
program type, credential type, program class, regustration status,
and gender.
}
\keyword{datasets}
